added sending faults to daughterboard CAN

+++ DistributedBMSDaughter/Drivers/DataHandlers/Inc/FaultManager.hpp
@@ -41,7 +41,7 @@
  
 private:
 
-     uint32_t fault_mask_;
+     uint8_t fault_mask_;
 
 
  };
